requirement.txt
transformers==4.31.0
datasets
deepspeed
sentencepiece
tensorboardX
accelerate

einops
importlib-resources
jsonlines
numexpr
omegaconf
pycountry
pytablewriter
rouge-score
sacrebleu
scikit-learn
sqlitedict
tqdm-multiprocess
zstandard